refactor MongoCollection read methods
